# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: PEPA — Platform Engineering & Pipeline Automator API
  description: |
    PEPA — Platform Engineering & Pipeline Automator API for managing services, workflows, scorecards,
    plugins, RBAC, and platform operations.
  version: "0.1.0"
  contact:
    name: PEPA Team
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8088/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /roles:
    get:
      tags: [RBAC]
      summary: List roles
      operationId: listRoles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Role"
                  total: { type: integer }
    post:
      tags: [RBAC]
      summary: Create role
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string }
                description: { type: string }
                scope: { type: string, default: tenant }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [RBAC]
      summary: Delete role
      operationId: deleteRole
      responses:
        "200":
          description: Deleted
  /roles/{id}/permissions:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [RBAC]
      summary: Get role permissions
      operationId: getRolePermissions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  permissions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Permission"
                  total: { type: integer }
    post:
      tags: [RBAC]
      summary: Add permission to role
      operationId: addRolePermission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resource, action]
              properties:
                resource: { type: string }
                action: { type: string }
                effect: { type: string, default: allow }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permission"
  /role-assignments:
    get:
      tags: [RBAC]
      summary: List role assignments
      operationId: listAssignments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  assignments:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoleAssignment"
                  total: { type: integer }
    post:
      tags: [RBAC]
      summary: Assign role to user
      operationId: assignRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, role_id]
              properties:
                user_id: { type: string, format: uuid }
                role_id: { type: string, format: uuid }
      responses:
        "201":
          description: Created
  /role-assignments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [RBAC]
      summary: Revoke role assignment
      operationId: revokeAssignment
      responses:
        "200":
          description: Assignment revoked
...
  /me/check:
    get:
      tags: [User]
      summary: Check current user's permission
      operationId: checkMyPermission
      parameters:
        - name: resource
          in: query
          required: true
          schema: { type: string }
        - name: action
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowed: { type: boolean }

# ── Components ──────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        is_system: { type: boolean }
        scope: { type: string }
    Permission:
      type: object
      properties:
        id: { type: string, format: uuid }
        role_id: { type: string, format: uuid }
        resource: { type: string }
        action: { type: string }
        effect: { type: string }
    RoleAssignment:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        role_id: { type: string, format: uuid }
        role_name: { type: string }
        is_active: { type: boolean }